// service/types.gen.ts
export type PostPositionAddData = {
    body?: {
        /**
         * 商圈ID
         */
        cbdId: string;
        /**
         * 分区ID
         */
        partId: string;
        /**
         * 铺位编号
         */
        no: string;
    };
    path?: never;
    query?: never;
    url: '/position/add';
};